requirements.txt
mido==1.3.3
numpy==2.2.0
packaging==24.2
pygame==2.6.1
PyOpenGL==3.1.7
